Rename hack module
